checkpoint for color detection

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/auto/CapstoneDetectionCamera.java
@@ -48,7 +48,7 @@ public class CapstoneDetectionCamera {
         return pipeline.position;
     }
 
-    public int[] getAnalysis() {
+    public double[] getAnalysis() {
         return pipeline.getAnalysis();
     }
 }
